Updated Tidy's selectors to work with the latest version of the module.
Updated item dataset retrieval to perform a `closest` check, to handle cases like Tidy's, where the row to colorize is not also the dataset node. It still works the same for other sheets.

Updated the background setting code to also set the `background` in addition to `background-color`, to prevent cases like backgrounds with gradients overlapping with the rarity color unintentionally.